+++ src/source/lib/config/event.ts
@@ -1,0 +1,38 @@
+export const EventOptions = [
+  {
+    label: '元素点击',
+    value: 'CLICK',
+  },
+  {
+    label: '表单聚焦',
+    value: 'FOCUS',
+  },
+  {
+    label: '表单输入',
+    value: 'INPUT',
+  },
+  {
+    label: '表单赋值',
+    value: 'CHANGE',
+  },
+  {
+    label: '表单失焦',
+    value: 'BLUR',
+  },
+  {
+    label: 'Enter按下',
+    value: 'KEY_DOWN',
+  },
+  {
+    label: 'Enter抬起',
+    value: 'KEY_UP',
+  },
+  {
+    label: '鼠标按下',
+    value: 'MOUSE_DOWN',
+  },
+  {
+    label: '鼠标抬起',
+    value: 'MOUSE_UP',
+  },
+];